CSHARP-2119: Test that killCursors really succeeds
