+++ tests/MongoDB.Driver.Core.Tests/Core/Operations/AsyncCursorTests.cs
@@ -454,46 +454,6 @@ namespace MongoDB.Driver.Core.Operations
             sameSessionWasUsed.Should().BeTrue();
         }
 
-        [Theory]
-        [ParameterAttributeData]
-        public void KillCursor_should_actually_work(
-            [Values(false, true)] bool async)
-        {
-            using (var client = DriverTestConfiguration.CreateDisposableClient())
-            {
-            var database = client.GetDatabase("admin");
-            var collection = database.GetCollection<BsonDocument>(GetType().Name);
-
-                if (async)
-                {
-                    // insert 1000 documents into a collection
-                    for (int i = 0; i < 1000; i++)
-                    {
-                        collection.InsertOneAsync(new BsonDocument("x", i)).GetAwaiter().GetResult();
-                    }
-
-                    var cursor = collection.FindAsync("{x: 2}").GetAwaiter().GetResult();
-
-                }
-                else
-                {
-                    // insert 1000 documents into a collection
-                    for (int i = 0; i < 1000; i++)
-                    {
-                        collection.InsertOne(new BsonDocument("x", i));
-                    }
-
-                    var cursor = collection.FindSync("{x: 2}");
-                }
-            }
-//            still need to do the following:
-//            - store the cursor id in a variable
-//            - assert the cursor id is nonzero
-//            - kill the cursor (explicitly, or by letting it go out of scope, depending on the driver)
-//            - assert that the driver receives a server reply to "killCursors" with "ok: 1", an empty "cursorsNotFound" array,
-//             an empty "cursorsAlive" array, an empty "cursorsUnknown" array, and a "cursorsKilled" array with one element equal to the cursor id
-        }
-
         // private methods
         private ConnectionDescription CreateConnectionDescriptionSupportingSession(string version = "3.6.0")
         {
