+++ tests/MongoDB.Driver.Core.Tests/Core/Operations/CursorTests.cs
@@ -17,7 +17,6 @@ using System.Threading;
 using FluentAssertions;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Serializers;
-using MongoDB.Driver.Builders;
 using MongoDB.Driver.Core.Bindings;
 using MongoDB.Driver.Tests;
 using Moq;
